add test for settings_parser
